# docs/09-api/openapi.yaml
openapi: 3.1.0
info:
  title: Lumine API
  description: Institutional AI-native quantitative hedge fund API
  version: 0.1.0
paths:
...
  /api/v1/market/bars:
    get:
      tags:
      - market
      summary: List Bars
      description: Return recent market bars.
      operationId: list_bars_api_v1_market_bars_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList_MarketBar_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/signals/{symbol}:
    get:
      tags:
      - market
      summary: List Symbol Signals
      description: Return recent analyst signals for one symbol (B-06).
      operationId: list_symbol_signals_api_v1_market_signals__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList_Signal_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/signals:
    get:
      tags:
      - market
      summary: List Signals
      description: Return recent analyst signals.
      operationId: list_signals_api_v1_market_signals_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList_Signal_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/quote/{symbol}:
    get:
      tags:
      - market
      summary: Get Quote
      description: Return the current bid/ask/last quote for a symbol.
      operationId: get_quote_api_v1_market_quote__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketQuote'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/quotes:
    get:
      tags:
      - market
      summary: Get Quotes
      description: Batch fetch quotes for multiple symbols (map of symbol → quote).
      operationId: get_quotes_api_v1_market_quotes_get
      parameters:
      - name: symbols
        in: query
        required: true
        schema:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 50
          title: Symbols
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/MarketQuote'
                title: Response Get Quotes Api V1 Market Quotes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/ohlcv/{symbol}:
    get:
      tags:
      - market
      summary: Get Ohlcv
      description: Return OHLCV bars for a symbol (deterministic demo random walk).
      operationId: get_ohlcv_api_v1_market_ohlcv__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: timeframe
        in: query
        required: false
        schema:
          type: string
          pattern: ^(1m|5m|15m|30m|1h|4h|1d|1w)$
          default: 1h
          title: Timeframe
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 10000
          minimum: 1
          default: 100
          title: Limit
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Since
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarketBar'
                title: Response Get Ohlcv Api V1 Market Ohlcv  Symbol  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/symbol/{symbol}:
    get:
      tags:
      - market
      summary: Get Symbol Config
      description: Return instrument specification for a symbol.
      operationId: get_symbol_config_api_v1_market_symbol__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolConfig'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/symbols:
    get:
      tags:
      - market
      summary: List Symbols
      description: List all active instruments (optionally including inactive ones).
      operationId: list_symbols_api_v1_market_symbols_get
      parameters:
      - name: asset_class
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset Class
      - name: exchange
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Exchange
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Inactive
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SymbolConfig'
                title: Response List Symbols Api V1 Market Symbols Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/volatility/{symbol}:
    get:
      tags:
      - market
      summary: Get Volatility
      description: Return rolling volatility (fraction) for a symbol.
      operationId: get_volatility_api_v1_market_volatility__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: window
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          default: 14
          title: Window
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolatilityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/correlation:
    get:
      tags:
      - market
      summary: Get Correlation
      description: Return a symmetric correlation matrix for the requested symbols.
      operationId: get_correlation_api_v1_market_correlation_get
      parameters:
      - name: symbols
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          title: Symbols
      - name: window
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          default: 30
          title: Window
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  additionalProperties:
                    type: number
                title: Response Get Correlation Api V1 Market Correlation Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/spread/{symbol}:
    get:
      tags:
      - market
      summary: Get Spread
      description: Return spread statistics for a symbol over a period.
      operationId: get_spread_api_v1_market_spread__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: period
        in: query
        required: false
        schema:
          type: integer
          maximum: 86400
          minimum: 1
          default: 60
          title: Period
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpreadMetrics'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/session/{symbol}:
    get:
      tags:
      - market
      summary: Get Session
      description: Return current trading session state for a symbol.
      operationId: get_session_api_v1_market_session__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/market/features/{symbol}:
    get:
      tags:
      - market
      summary: Get Features
      description: Return computed technical features for a symbol (FeaturePanel contract).
      operationId: get_features_api_v1_market_features__symbol__get
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureSet'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MarketBar:
      properties:
        symbol:
          type: string
          title: Symbol
        timeframe:
          type: string
          title: Timeframe
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        open:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Open
        high:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: High
        low:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Low
        close:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Close
        volume:
          type: integer
          title: Volume
      type: object
      required:
      - symbol
      - timeframe
      - timestamp
      - open
      - high
      - low
      - close
      - volume
      title: MarketBar
      description: OHLCV bar for market data endpoints.
    MarketQuote:
      properties:
        symbol:
          type: string
          title: Symbol
        bid:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Bid
        ask:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Ask
        mid:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Mid
        last:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Last
        volume_24h:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Volume 24H
        change_24h:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Change 24H
        change_pct_24h:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Change Pct 24H
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
      - symbol
      - bid
      - ask
      - mid
      - last
      - volume_24h
      - change_24h
      - change_pct_24h
      - timestamp
      title: MarketQuote
      description: Current bid/ask/last snapshot for a symbol.
...
    PaginatedList_MarketBar_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MarketBar'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PaginatedList[MarketBar]
...
    PaginatedList_Signal_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Signal'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PaginatedList[Signal]
...
    VolatilityResponse:
      properties:
        volatility:
          type: number
          title: Volatility
      type: object
      required:
      - volatility
      title: VolatilityResponse
      description: Rolling volatility for a symbol (fraction, e.g. 0.12 = 12%).
...